requirements.txt
pytube
python-dateutil
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
